requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
qrcode[pil]==7.4.2
Pillow==10.4.0
python-dotenv==1.0.0
requests==2.31.0
fpdf2==2.7.8
